Refactor: Combine sphinxCanProcess file from its two implementations.
